Fixed unreasonable inputs and separated DataStract from utility structs
