+++ petrov.konstantin/T2/DataStruct.cpp
@@ -3,6 +3,7 @@
 #include <string>
 #include <bitset>
 #include <cstring>
+#include "FormatUtils.h"
 
 bool petrov::operator<(const DataStruct& left, const DataStruct& right)
 {
@@ -43,10 +44,10 @@ std::istream& petrov::operator>>(std::istream& in, DataStruct& dest)
   }
   DataStruct input;
   {
-    using sep = DelimiterI;
-    using sllLit = SignedLongLongLiteralI;
-    using ullBin = UnsignedLongLongBinaryI;
-    using str = StringI;
+    using sep = petrov::DelimiterI;
+    using sllLit = petrov::SignedLongLongLiteralI;
+    using ullBin = petrov::UnsignedLongLongBinaryI;
+    using str = petrov::StringI;
     in >> sep{ '(' } >> sep{ ':' };
     std::string key = "";
     for (size_t i = 0; i < 3; ++i)
@@ -75,84 +76,6 @@ std::istream& petrov::operator>>(std::istream& in, DataStruct& dest)
   if (in)
   {
     dest = input;
-  }
-  return in;
-}
-std::istream& petrov::operator>>(std::istream& in, SignedLongLongLiteralI&& dest)
-{
-  std::istream::sentry sentry(in);
-  if (!sentry)
-  {
-    return in;
-  }
-  in >> dest.ref;
-  in >> DelimiterI{ 'l' } >> DelimiterI{ 'l' };
-  return in;
-}
-std::istream& petrov::operator>>(std::istream& in, UnsignedLongLongBinaryI&& dest)
-{
-  std::istream::sentry sentry(in);
-  if (!sentry)
-  {
-    return in;
-  }
-  in >> DelimiterI{ '0' };
-  char c = '0';
-  if ((in >> c) && (c != 'b') && (c != 'B'))
-  {
-    in.setstate(std::ios::failbit);
-  }
-  else if (in)
-  {
-    char binary[64] = "";
-    in.get(binary, 64, ':');
-    for (size_t i = 0; binary[i] != '\0'; ++i)
-    {
-      if (!std::isdigit(binary[i]))
-      {
-        in.setstate(std::ios::failbit);
-        return in;
-      }
-    }
-    dest.ref = std::stoull(binary, nullptr, 2);
-  }
-  return in;
-}
-std::istream& petrov::operator>>(std::istream& in, StringI&& dest)
-{
-  std::istream::sentry sentry(in);
-  if (!sentry)
-  {
-    return in;
-  }
-  return std::getline(in >> DelimiterI{ '\"' }, dest.ref, '\"');
-}
-std::istream& petrov::operator>>(std::istream& in, DelimiterI&& dest)
-{
-  std::istream::sentry sentry(in);
-  if (!sentry)
-  {
-    return in;
-  }
-  char c = '0';
-  in >> c;
-  if (in && (c != dest.expected))
-  {
-    in.setstate(std::ios::failbit);
-  }
-  return in;
-}
-std::istream& petrov::operator>>(std::istream& in, LabelI&& dest)
-{
-  std::istream::sentry sentry(in);
-  if (!sentry)
-  {
-    return in;
-  }
-  std::string data = "";
-  if ((in >> data) && (data != dest.expected))
-  {
-    in.setstate(std::ios::failbit);
   }
   return in;
 }
